brcmsmac: use read after write workaround only for PCIe devices
Accordingly to the comment over this code, this workaround is just
needed on a PCIe devices connected to a broken PCIe host controller
found in the BCM4716 and BCM4706 SoC. This patches changes the code in
a way that it is only activated on such configurations. This was tested
on a BCM4716 with the wifi core on the SoC and the wifi core on the
card though the broken PCIe controller.