MIPS: Use same definition for tlbmiss_handler_setup_pgd
tlbmiss_handler_setup_pgd is defined as a pointer to a u32 array in
tlbex.c and as a function pointer in mmu_context.h. This was done
because tlbex.c fills the memory of u32 with assembler code which
implements a function, this assembler code depends on the CPU being
used. Later the code will jump into this function.

This patch uses the same type for both definitions and makes use of the
pointer to the _start of the function in places where we have to access
the code of the function.

This fixes the build with LTO.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>